# cmd/api/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a Fiber application.",
        "title": "Fiber Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Ryudelta Support",
            "url": "https://t.me/Grafiters",
            "email": "ryudelta7@gmail.com"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
...
        "/api/transaction/create": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Create Transaction data for non crucial data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Transaction",
                "parameters": [
                    {
                        "description": "transaction to create",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.TransactionInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.SingleResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.TransactionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/domain.SingleResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.SingleResponse"
                        }
                    }
                }
            }
        },
        "/api/transaction/get": {
            "get": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Get Transaction data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Get Transaction",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.PaginationResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.TransactionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/domain.SingleResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.SingleResponse"
                        }
                    }
                }
            }
        },
        "/api/transaction/get/{id}": {
            "get": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Get Transaction data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Get Transaction",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.PaginationResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.TransactionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/domain.SingleResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.SingleResponse"
                        }
                    }
                }
            }
        },
        "/api/transaction/installment/pay": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Create Transaction data for non crucial data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Transaction",
                "parameters": [
                    {
                        "description": "transaction to pay installment",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.BulkInstallmentUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.SingleResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.TransactionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/domain.SingleResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.SingleResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.BulkInstallmentUpdate": {
            "type": "object",
            "properties": {
                "installment_update": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.InstallmentUpdae"
                    }
                },
                "transaction_id": {
                    "type": "integer"
                }
            }
        },
        "domain.CustomerResponse": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "birth_place": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ktp_image_url": {
                    "type": "string"
                },
                "legal_name": {
                    "type": "string"
                },
                "limit": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.LimitResponse"
                    }
                },
                "nik": {
                    "type": "integer"
                },
                "salary": {
                    "type": "integer"
                },
                "selfie_image_url": {
                    "type": "string"
                }
            }
        },
...
        "domain.InstallmentLogRespponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "month": {
                    "type": "integer"
                },
                "paid_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.InstallmentUpdae": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "due_date": {
                    "type": "string"
                },
                "month": {
                    "type": "integer"
                },
                "paid_at": {
                    "type": "string"
                }
            }
        },
        "domain.LimitResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "tenors": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.PaginationResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "number",
                    "type": "integer"
                },
                "data": {
                    "description": "list of data"
                },
                "message": {
                    "description": "string",
                    "type": "string"
                }
            }
        },
...
        "domain.SingleResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "number",
                    "type": "integer"
                },
                "data": {
                    "description": "single data"
                },
                "message": {
                    "description": "string",
                    "type": "string"
                }
            }
        },
...
        "domain.TransactionInput": {
            "type": "object",
            "properties": {
                "admin_fee": {
                    "type": "integer"
                },
                "asset_name": {
                    "type": "string"
                },
                "contract_no": {
                    "type": "integer"
                },
                "installment": {
                    "type": "integer"
                },
                "otr": {
                    "type": "integer"
                },
                "tenor": {
                    "type": "integer"
                }
            }
        },
        "domain.TransactionResponse": {
            "type": "object",
            "properties": {
                "admin_fee": {
                    "type": "integer"
                },
                "asset_name": {
                    "type": "string"
                },
                "contract_no": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "customers": {
                    "$ref": "#/definitions/domain.CustomerResponse"
                },
                "id": {
                    "type": "integer"
                },
                "installment": {
                    "type": "integer"
                },
                "installment_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.InstallmentLogRespponse"
                    }
                },
                "limits": {
                    "$ref": "#/definitions/domain.LimitResponse"
                },
                "otr": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Token": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}